// src/modules/message/enums/message.enums.ts
export enum MessageStatusEnum {
  PENDING = 'pending',
  ACCEPTED = 'accepted',
  REJECTED = 'rejected',
  SENT = 'sent',
}

export enum MessageFolderEnum {
  INBOX = 'inbox',
  SENT = 'sent',
  DRAFTS = 'drafts',
  REQUESTS = 'requests',
}

export enum ProfileTypeEnum {
  STARTUP = 'startup',
  INVESTOR = 'investor',
}